chore: 🤖 rename typography components
