+++ sapper/src/components/typography/index.js
@@ -1,6 +1,6 @@
-import BaseParagraph from './BaseParagraph.svelte'
-import LeadingParagraph from './LeadingParagraph.svelte'
+import ParagraphSm from './ParagraphSm.svelte'
+import ParagraphMd from './ParagraphMd.svelte'
 import H1 from './H1.svelte'
 import H4 from './H4.svelte'
 
-export default { BaseParagraph, H1, H4, LeadingParagraph }
+export default { ParagraphSm, H1, H4, ParagraphMd }
